Develop/Ecp.True/DI/Ecp.True.DI.Analytics/requirements.txt
Click==7.0
Flask==1.1.1
flask_restplus==0.13.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
pandas==0.25.3
numpy==1.17.4
joblib==0.14.1
pyodbc==4.0.27
pmdarima==1.5.3
portalocker==1.5.2
statsmodels==0.11.0
holidays==0.9.11
xgboost==0.90
fbprophet==0.5.0
azure-common==1.1.23
azure-core==1.1.1
azure-keyvault==4.0.0
azure-keyvault-keys==4.0.0
azure-keyvault-secrets==4.0.0
azure-mgmt-resource==7.0.0
azure-mgmt-sql==0.15.0
azure-identity==1.1.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==1.5.0
azure-storage-file==1.4.0
azure-storage-nspkg==3.1.0
azure-storage-queue==1.4.0